pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "importlib-resources"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-human-inventory-game"
version = "1.0.0"
description = "Fullstack web application for exploring human-AI collaboration in inventory management"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name="Tianyi Peng", email="tianyipeng@gmail.com"},
    {name="Jackie Baek"},
    {name="Yaopeng Fu"},
    {name="Will Ma"},
]
keywords = ["inventory", "management", "ai", "llm", "operations-research", "education"]
dependencies = [
    "openai>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "supabase>=2.0.0",
    "httpx>=0.24.0",
    "pandas>=1.5.0",
    "python-dotenv>=1.0.0",
    "PyJWT>=2.8.0",
    "scipy>=1.11.0",
]

[project.urls]
"Homepage" = "https://github.com/TianyiPeng/ai-human-inventory-game"
"Documentation" = "https://github.com/TianyiPeng/ai-human-inventory-game"
"Repository" = "https://github.com/TianyiPeng/ai-human-inventory-game"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21.0", "black>=23.0", "flake8>=6.0"]

# Explicit setuptools configuration to avoid packaging unnecessary folders
[tool.setuptools.packages.find]
include = ["textarena*"]
exclude = ["data*", "tests*", "examples*"]

